requirements.txt
# k-Wave Python bindings
k-wave-python

# Core scientific computing
numpy>=1.20.0
scipy>=1.7.0
h5py>=3.0.0

# Medical image processing
pydicom>=2.3.0
nibabel>=4.0.0  # For NIfTI export

# Visualization
matplotlib>=3.5.0
pillow>=9.0.0  # For GIF generation
vtk>=9.0.0  # For 3D visualization
# Optional: For MP4 movie generation (requires ffmpeg installed separately)
# ffmpeg-python>=0.2.0
